feat: middleware for authUser

+++ src/app/api/admin/delete-product/route.js
@@ -1,4 +1,5 @@
 import dbConnection from "@/database";
+import AuthUser from "@/middleware/AuthUser";
 import Product from "@/models/product";
 import { NextResponse } from "next/server";
 
@@ -8,27 +9,36 @@ export async function DELETE(request) {
   try {
     await dbConnection();
 
-    const { searchParams } = new URL(request.url);
-    const id = searchParams.get("id");
+    const isAuthUser = await AuthUser(request);
 
-    if (!id) {
-      return NextResponse.json({
-        success: false,
-        message: "Product ID is required",
-      });
-    }
+    if (isAuthUser?.role === "admin") {
+      const { searchParams } = new URL(request.url);
+      const id = searchParams.get("id");
 
-    const deletedProduct = await Product.findByIdAndDelete(id);
+      if (!id) {
+        return NextResponse.json({
+          success: false,
+          message: "Product ID is required",
+        });
+      }
 
-    if (deletedProduct) {
-      return NextResponse.json({
-        success: true,
-        message: "Product deleted successfully!",
-      });
+      const deletedProduct = await Product.findByIdAndDelete(id);
+
+      if (deletedProduct) {
+        return NextResponse.json({
+          success: true,
+          message: "Product deleted successfully!",
+        });
+      } else {
+        return NextResponse.json({
+          success: false,
+          message: "Failed to delete product, please try again.",
+        });
+      }
     } else {
       return NextResponse.json({
         success: false,
-        message: "Failed to delete product, please try again.",
+        message: "You are not Authorized delete!",
       });
     }
   } catch (err) {
